backend/requirement.txt

Flask==2.3.2
pymongo==4.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
flask_cors